requirements.txt
fastapi[all]>=0.109.0
uvicorn[standard]>=0.27.0
SQLAlchemy>=2.0.0
asyncpg>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
